Extract Utf8 oop cache assertions into support module
expectUtf8OopToResolveViaSymbolLookup and expectUtf8OopToBeCached were
private to gciLibrary.test.ts's describe. A second suite now needs the
same assertion, and a copy would have dropped the reasoning the first
one carries -- the looked-up session is checked with toBe rather than
toHaveBeenCalledWith because koffi pointers have no enumerable
properties, so vitest's deep equality treats any two sessions as equal.

Both move together, along with the spyOnResolveSymbol wrapper they
share, which stays private to the new module. The sessionToUse default
goes away with the enclosing closure, so call sites now name the
session and the library they observe.
